% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_to_table.R
\name{db_summary}
\alias{db_summary}
\title{Quick database summary}
\usage{
db_summary(db)
}
\arguments{
\item{db}{an RDatabase object}
}
\value{
invisibly returns a data.frame with the summary
}
\description{
Print a summary of database contents by species and gene
}
