% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{filter_db}
\alias{filter_db}
\title{Filter database via species/gene/score and return a new RDatabase}
\usage{
filter_db(db, species = NULL, gene = NULL, min_vdjdb_score = 0L)
}
\arguments{
\item{db}{an RDatabase object}

\item{species}{optional species filter (e.g. "HomoSapiens")}

\item{gene}{optional gene filter (e.g. "TRB")}

\item{min_vdjdb_score}{minimum VDJdb score (0-3)}
}
\value{
an RDatabase object
}
\description{
Filter database via species/gene/score and return a new RDatabase
}
