% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{filter_db_by_epitope_size}
\alias{filter_db_by_epitope_size}
\title{Filter by minimum epitope size (unique CDR3 per epitope).}
\usage{
filter_db_by_epitope_size(db, min_size)
}
\description{
Filter by minimum epitope size (unique CDR3 per epitope).
}
