% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{match_tcr}
\alias{match_tcr}
\title{Match a single clonotype against the database.
Returns a list of columns (vector-of-equal-length) suitable for as.data.frame in R.}
\usage{
match_tcr(db, cdr3, v_segment, j_segment, scope, top_n)
}
\description{
Match a single clonotype against the database.
Returns a list of columns (vector-of-equal-length) suitable for as.data.frame in R.
}
