% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{vdjdb_download}
\alias{vdjdb_download}
\title{Ensure VDJdb exists locally; return the file path}
\usage{
vdjdb_download(use_fat_db = FALSE)
}
\arguments{
\item{use_fat_db}{logical; TRUE for full db, FALSE for slim (default)}
}
\value{
string file path
}
\description{
Ensure VDJdb exists locally; return the file path
}
