% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{vdjdb_ensure_into}
\alias{vdjdb_ensure_into}
\title{Ensure VDJdb exists in the specified directory and return the path.}
\usage{
vdjdb_ensure_into(dir, use_fat_db)
}
\description{
Ensure VDJdb exists in the specified directory and return the path.
}
