% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjdb.R
\name{vdjdb_packaged_path}
\alias{vdjdb_packaged_path}
\title{Locate packaged VDJdb file bundled with vdjmatchR}
\usage{
vdjdb_packaged_path(use_fat_db = FALSE)
}
\arguments{
\item{use_fat_db}{logical; TRUE for full db if bundled, FALSE for slim (default)}
}
\value{
string file path (empty string if not found)
}
\description{
Locate packaged VDJdb file bundled with vdjmatchR
}
