% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjdb.R
\name{vdjdb_path}
\alias{vdjdb_path}
\title{Preferred VDJdb path (user-specified if set, else packaged)}
\usage{
vdjdb_path(use_fat_db = FALSE)
}
\arguments{
\item{use_fat_db}{logical; TRUE for full db, FALSE for slim (default)}
}
\value{
string file path
}
\description{
Returns a user-specified path if configured via \link[=vdjdb_set_user_db]{vdjdb_set_user_db},
otherwise returns the packaged file path.
}
