% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjdb.R
\name{vdjdb_update_all}
\alias{vdjdb_update_all}
\title{Update/download both slim and fat VDJdb files (disabled)}
\usage{
vdjdb_update_all()
}
\description{
Update/download both slim and fat VDJdb files (disabled)
}
