% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{vdjdb_update_into}
\alias{vdjdb_update_into}
\title{Download/update the VDJdb files (slim and fat) into the specified directory.}
\usage{
vdjdb_update_into(dir)
}
\description{
Download/update the VDJdb files (slim and fat) into the specified directory.
}
