% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdjdb.R
\name{vdjdb_update_latest}
\alias{vdjdb_update_latest}
\title{Update (download) the VDJdb file and return its path into extdata}
\usage{
vdjdb_update_latest(use_fat_db = FALSE)
}
\arguments{
\item{use_fat_db}{logical; TRUE for full db, FALSE for slim (default)}
}
\value{
string file path to the updated file
}
\description{
Update (download) the VDJdb file and return its path into extdata
}
