% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{vdjmatchR-package}
\alias{vdjmatchR}
\alias{vdjmatchR-package}
\title{vdjmatchR: R Interface to VDJmatch (Rust)}
\description{
R bindings to query TCR sequences using a Rust implementation derived from vdjmatch-rs, via extendr.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yourname/vdjmatchR}
  \item Report bugs at \url{https://github.com/yourname/vdjmatchR/issues}
}

}
\author{
\strong{Maintainer}: vdjmatch Authors \email{you@example.com}

}
