% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{filter_db}
\alias{filter_db}
\title{Filter database entries by species, gene, and minimum VDJdb score.}
\usage{
filter_db(db, species, gene, min_vdjdb_score)
}
\description{
Filter database entries by species, gene, and minimum VDJdb score.
}
