% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{filter_db_by_epitope_size}
\alias{filter_db_by_epitope_size}
\title{Filter database by minimum epitope size}
\usage{
filter_db_by_epitope_size(db, min_size = 1L)
}
\arguments{
\item{db}{an RDatabase object}

\item{min_size}{minimum number of unique CDR3s per epitope}
}
\value{
an RDatabase object
}
\description{
Filter database by minimum epitope size
}
