% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/package.R
\docType{package}
\name{vdjmatchR-package}
\alias{vdjmatchR-package}
\title{vdjmatchR: R Interface to VDJmatch (Rust)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R bindings to query TCR sequences using a Rust implementation derived from vdjmatch-rs, via extendr.

Provides fast TCR matching against VDJdb-formatted databases using a Rust
engine exposed via extendr. Bundles slim/full sample databases and supports
using user-supplied database files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://furlan-lab.github.io/vdjmatchR}
  \item \url{https://github.com/furlan-lab/vdjmatchR}
  \item Report bugs at \url{https://github.com/furlan-lab/vdjmatchR/issues}
}


Useful links:
\itemize{
  \item \url{https://furlan-lab.github.io/vdjmatchR}
  \item \url{https://github.com/furlan-lab/vdjmatchR}
  \item Report bugs at \url{https://github.com/furlan-lab/vdjmatchR/issues}
}

}
\author{
\strong{Maintainer}: vdjmatch Authors \email{you@example.com}

}
\keyword{internal}
